requirements.txt
openai>=2.8.1
faiss-cpu>=1.13.0
pydantic>=2.12.5
numpy>=2.3.5
fastapi>=0.123.10
uvicorn>=0.30.0
python-dotenv>=1.2.1